Add ShowRooms functionality
Adds a componentDidMount function to MainContent to check the database for any existing rooms upon loading the page. Only checks for existing rooms if none pop up in the state, because it assumes this would have already been run before state was populated.

If rooms exist, it checks to see if the current user is a valid user, and if so, adds that room to the state. It does this for all applicable rooms. Once the state gets updated, the componentDidUpdate function will create the JSX for the nav bar and render the page to include it.

Tested against rooms that the user is not a valid user of and it passes all tests.